requirements.txt
bcrypt==3.2.0
cffi==1.14.6
click==7.1.2
cryptography==3.4.8
ecdsa==0.17.0
fastapi==0.65.0
greenlet==1.1.0
h11==0.12.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.1
pyodbc==4.0.32
python-dotenv==0.19.0
python-jose==3.3.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.15
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.13.4
